New games marked as GAME_NOT_WORKING ------------------------------------ Rapid River
Please credit Guru for his dump and his readme.

The ROMs and readme were put in namcos23.c since they fit well, but the set doesn't get as far as the rest of the games in the driver.
Roms moved from /roms/unemulated/Namco/system23 to /roms/current/r.
Updated the GAME_NOT_WORKING page on the Wiki.