From: Fabio Priuli [mailto:doge.fabio@gmail.com] Subject: more patches Hi,
enclosed please find the following patches:

spaceinv.diff : this is an important patch. Please revert (if applied) 
my previous fix to bugs 2143 & 2144. It was wrong. This is the correct 
fix for those two bugs and many other issues not reported on MT (e.g. 
spaceatt, spaceat2 not starting because tilted, missing dips in many 
invaders clones etc). Basically, in 8080bw.c all the "DSW0" input ports 
should have been "IN2" for invaders_map to work correctly, and no 
additional "IN2" is necessary. After this fix, you may want to change 
all the "IN2" occurrences 8080bw.c and in mw8080bw.c with "DSW" since 
they mostly are dips. But the latter is a cosmetic change, so it's up 
to you decide if it's worth or not (the patch leaves "IN2").

avgdvg.diff : this correctly fixes avgdvg_done usage without using a 
parameter, but using IPT_ACTIVE_LOW where needed (I'm a bit ashamed 
because I should have thought to that since the beginning).

galaxold.diff : this changes ckongg coinage inputs to fix bug 2154. 
Notice anyway that coinage settings in dips do not correspond to the 
numbers which appear on screen. This is probably due to the conversion 
to galaxian hardware (e.g. it's shown the value of coin2, but in ckongg 
there is no coin2 at all, so the appearance of the coin 2 value has no 
meaning). Notice also that an unused dips sets the possibility to play 
in 2 or not. I defaulted this to 1 player because setting "2 players" 
automatically gives 2/4 credits with a single coin, which I don't think 
it was the default back then ;) 

However, perhaps ckongg is worth further analysis when it will be 
updated to the new galaxian driver.

stv.diff : this patch fixes the longstanding bug 650 (duplicated P1 
button 4). After testing all sets which reach the test mode, I found 
out that only batmanfr uses the button4 in port "P1", while all other 
games with more than 3 buttons use the one in "EXTRA". Therefore I made 
the "P1" 4th buttons a specific batmanfr input.

patch01.diff, patch02.diff, sys1.diff : these patches simplifies input 
ports in the remaining driver starting with S (plus a few small changes 
to other drivers I forgot previously), using PORT_INCLUDE, AM_READ_PORT 
& converting 'fake' 8bit ports to full 16/32bit ports  where suitable

Let me know if something fails to apply, or if there is any other problem