[Pugsy]
There's an issue with the debugger handling of the 6502 instructions as
there's a mismatch between the core and what the debugger shows.
The changes are detailed in the spreadsheets that I've attached that
compares the core to the dasm.


Attachments:

m6502fix.zip contains the diff file
6502 debugger disassembly changes.zip   contains an excel spreadsheet
(in both xls and xml formats) detailing the reasons/changes.