from MAMEPlus
this documents the CPS2 Phoenix sets, and what happens to a dead CPS2 board.

I think it's important we start keeping track of these now as they're finding their way onto bootleg boards, with further hacks and modifications made along the way.

Far from an ideal situation, but, our job is to document the facts, not discriminate by ignoring them.