NOTE: With the change, all existing CHD diff files are invalid.
Updated CHD format to version 4. Checksums are now computed and
stored separately for raw data and metadata, and metadata is now
checksummed by default.

We will need to go through all existing CHDs, run a chdman -update
on them, and update the SHA1s stored in the drivers to accommodate
this (MD5s should be removed).

Updated chdman to support a generic metadata addition system:

  chdman -addmeta <chdfile> <tag> [<index>] <sourcefile>

The <sourcefile> is examined and if it appears to be strictly text,
any EOFs and trailing EOLs are stripped, and the result is NULL-
terminated to match the behavior of existing metadata.

Updated chdman to report and fix errors in the raw and metadata
SHA1s.

Changed the CHD verify interfaces to pass back a structure containing
all the necessary data for verification and fixing.