Eeprom and i2c implemeted as devices [Samuele Zannoli]
The eepromdev.* files are the device impementation for the serial eeproms
The i2cmemdev.* files are the device impementation for the i2c memory
These aren't currently hooked up

I'm committing this so that the Naomi changes don't accidentally get applied.  The idea seems good, but they conflict with the current driver and I'd rather not have them accidentally slipping in and breaking things just yet.