fixed mantis 3126
sbishi is a 2 player set, it uses the 'player 3' inputs for the 2nd player.

attempting to use the inputs mapped as 'player 2' in the parent set causes the game to break in strange ways, treating that player as both player 1 and/or player 2 depending on the situation, not awarding any score etc.

this removes the broken inputs for the 2 player set and renumbers the players correctly, as they are clearly not meant to be connected in this case.