From: Justin Kerk [mailto:dopefishjustin@gmail.com] Sent: Saturday, May 30, 2009 8:30 PM To: submit@mamedev.org Subject: Amiga Custom sound - Paula?
As part of the conversion of custom sound devices, the sound chip for
Amiga-hardware games is now named "Amiga Custom" on the Game
Information display. The Amiga sound chip was called Paula, so it
would be nicer to show that instead, as in the attached patch. (You
could also use "MOS 8364" but that's not as catchy....)

-Justin

-- 
dopefish justin at gmail dot com
http://interbutt.com/